+++ map2.c
@@ -13,7 +13,7 @@ void *map_get_elem(map_t *map, void *key, key_comparator_t key_cmp)
 {
     if (key == NULL)
         return NULL;
-    for (int i = 0; i < list_get_size(map); i++) {
+    for (int i = 0; i < (int) list_get_size(map); i++) {
         pair_t *pair = list_get_elem_at_position(map, i);
         if (key_cmp(key, pair->key) == 0)
             return pair->value;
